refactor account info as a separate module
